requirements.txt
alembic==1.9.1
attrs==22.2.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
importlib-metadata==5.2.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
numpy==1.24.1
packaging==22.0
pandas==1.5.2
pluggy==1.0.0
psycopg2-binary==2.9.5
pyodbc==4.0.35
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.45
tomli==2.0.1
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
python-dotenv==0.21.0